Refactor allocator for global svc deployment
- Move SimpleAllocator and AbstractAllocator into AllocatorServiceImpl
- Send event to external scheduler when deploying global service
