+++ code/iaas/allocator/pom.xml
@@ -28,5 +28,10 @@
             <artifactId>cattle-framework-jooq</artifactId>
             <version>${project.version}</version>
         </dependency>
+        <dependency>
+            <groupId>io.cattle</groupId>
+            <artifactId>cattle-iaas-agent-instance</artifactId>
+            <version>${project.version}</version>
+        </dependency>
     </dependencies>
 </project>
