+++ code/iaas/allocator/src/main/java/io/cattle/platform/allocator/dao/impl/AllocationCandidateIterator.java
@@ -1,4 +1,4 @@
-package io.cattle.platform.simple.allocator.dao.impl;
+package io.cattle.platform.allocator.dao.impl;
 
 import io.cattle.platform.allocator.service.AllocationCandidate;
 import io.cattle.platform.object.ObjectManager;
