+++ code/iaas/allocator/src/main/java/io/cattle/platform/allocator/dao/impl/AllocatorDaoImpl.java
@@ -39,6 +39,7 @@ import io.cattle.platform.core.model.Volume;
 import io.cattle.platform.core.model.VolumeStoragePoolMap;
 import io.cattle.platform.core.model.tables.records.HostRecord;
 import io.cattle.platform.core.model.tables.records.InstanceRecord;
+import io.cattle.platform.core.model.tables.records.PortRecord;
 import io.cattle.platform.core.model.tables.records.StoragePoolRecord;
 import io.cattle.platform.db.jooq.dao.impl.AbstractJooqDao;
 import io.cattle.platform.object.ObjectManager;
@@ -48,8 +49,10 @@ import io.github.ibuildthecloud.gdapi.condition.ConditionType;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -62,7 +65,10 @@ import org.jooq.Field;
 import org.jooq.Record;
 import org.jooq.Record3;
 import org.jooq.RecordHandler;
+import org.jooq.Result;
+import org.jooq.SelectConditionStep;
 import org.jooq.exception.InvalidResultException;
+import org.jooq.impl.DSL;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -76,13 +82,20 @@ public class AllocatorDaoImpl extends AbstractJooqDao implements AllocatorDao {
     private static final String PUBLIC_PORT = "publicPort";
     private static final String INSTANCE_ID = "instanceID";
     private static final String ALLOCATED_IPS = "allocatedIPs";
-    
     private static final String BIND_ADDRESS = "bindAddress";
+
+    private static final List<Field<?>> hostAndPortFields;
+    static {
+       hostAndPortFields = new ArrayList<Field<?>>(Arrays.asList(PORT.fields()));
+       hostAndPortFields.add(INSTANCE_HOST_MAP.HOST_ID);
+    }
 
     static final List<String> IHM_STATES = Arrays.asList(new String[] { CommonStatesConstants.INACTIVE, CommonStatesConstants.DEACTIVATING,
             CommonStatesConstants.REMOVED, CommonStatesConstants.REMOVING, CommonStatesConstants.PURGING, CommonStatesConstants.PURGED });
 
+    @Inject
     ObjectManager objectManager;
+    @Inject
     GenericMapDao mapDao;
 
     @Override
@@ -240,7 +253,7 @@ public class AllocatorDaoImpl extends AbstractJooqDao implements AllocatorDao {
         Map<Object, Object> criteria = new HashMap<Object, Object>();
         criteria.put(STORAGE_DRIVER.REMOVED, new io.github.ibuildthecloud.gdapi.condition.Condition(ConditionType.NULL));
         criteria.put(STORAGE_DRIVER.ID, new io.github.ibuildthecloud.gdapi.condition.Condition(ConditionType.IN, storageDriverIds));
-        List<StorageDriver> drivers = getObjectManager().find(StorageDriver.class, criteria);
+        List<StorageDriver> drivers = objectManager.find(StorageDriver.class, criteria);
         Map<Long, StorageDriver> storageDrivers = new HashMap<>();
         for (StorageDriver d : drivers) {
             storageDrivers.put(d.getId(), d);
@@ -369,24 +382,6 @@ public class AllocatorDaoImpl extends AbstractJooqDao implements AllocatorDao {
                     .and(INSTANCE_HOST_MAP.STATE.notIn(IHM_STATES))
                     .and((INSTANCE.HEALTH_STATE.isNull().or(INSTANCE.HEALTH_STATE.eq(HealthcheckConstants.HEALTH_STATE_HEALTHY)))))
                 .fetch().size() > 0;
-    }
-
-    public ObjectManager getObjectManager() {
-        return objectManager;
-    }
-
-    @Inject
-    public void setObjectManager(ObjectManager objectManager) {
-        this.objectManager = objectManager;
-    }
-
-    public GenericMapDao getMapDao() {
-        return mapDao;
-    }
-
-    @Inject
-    public void setMapDao(GenericMapDao mapDao) {
-        this.mapDao = mapDao;
     }
 
     @Override
@@ -563,5 +558,148 @@ public class AllocatorDaoImpl extends AbstractJooqDao implements AllocatorDao {
             }
         }
         return;
+    }
+    
+
+    @Override
+    public Iterator<AllocationCandidate> iteratorHosts(List<String> orderedHostUuids, List<Long> volumes, QueryOptions options) {
+        List<CandidateHostInfo> hostInfos = new ArrayList<>();
+        Set<Long> hostIds = new HashSet<>();
+        if (orderedHostUuids == null) {
+            Result<Record3<String, Long, Long>> result = getHostQuery(null, options).fetch();
+            Collections.shuffle(result);
+
+            Map<Long, CandidateHostInfo> infoMap = new HashMap<>();
+            for (Record3<String, Long, Long> r : result) {
+                Long hostId = r.value2();
+                hostIds.add(hostId);
+                CandidateHostInfo hostInfo = infoMap.get(hostId);
+                if (hostInfo == null) {
+                    hostInfo = new CandidateHostInfo(hostId, r.value1());
+                    infoMap.put(hostId, hostInfo);
+                    hostInfos.add(hostInfo);
+                }
+                hostInfo.getPoolIds().add(r.value3());
+            }
+        } else {
+            Map<String, Result<Record3<String, Long, Long>>> result = getHostQuery(orderedHostUuids, options).fetchGroups(HOST.UUID);
+            for (String uuid : orderedHostUuids) {
+                Result<Record3<String, Long, Long>>val = result.get(uuid);
+                if (val != null) {
+                    Set<Long>poolIds = new HashSet<>();
+                    Long hostId = null;
+                    for (Record3<String, Long, Long>r : val) {
+                        poolIds.add(r.value3());
+                        // host id is same in all records for this group
+                        if (hostId == null) {
+                            hostId = r.value2();
+                        }
+                    }
+
+                    CandidateHostInfo hostInfo = new CandidateHostInfo(hostId, uuid);
+                    hostInfo.getPoolIds().addAll(poolIds);
+                    hostInfos.add(hostInfo);
+                    hostIds.add(hostId);
+                }
+            }
+        }
+
+        if (options.isIncludeUsedPorts()) {
+            updateHostsWithUsedPorts(hostIds, hostInfos);
+        }
+
+        return new AllocationCandidateIterator(objectManager, hostInfos, volumes);
+    }
+
+    private void updateHostsWithUsedPorts(Set<Long> hostIds, List<CandidateHostInfo> hostInfos) {
+        Map<Long, Result<Record>> results = create()
+                .select(hostAndPortFields)
+                    .from(PORT)
+                    .join(INSTANCE_HOST_MAP)
+                        .on(PORT.INSTANCE_ID.eq(INSTANCE_HOST_MAP.INSTANCE_ID))
+                    .join(INSTANCE)
+                        .on(INSTANCE_HOST_MAP.INSTANCE_ID.eq(INSTANCE.ID))
+                .leftOuterJoin(SERVICE_EXPOSE_MAP)
+                .on(SERVICE_EXPOSE_MAP.INSTANCE_ID.eq(INSTANCE.ID))
+                    .where(INSTANCE_HOST_MAP.HOST_ID.in(hostIds)
+                        .and(INSTANCE.REMOVED.isNull())
+                        .and(INSTANCE.STATE.in(InstanceConstants.STATE_STARTING, InstanceConstants.STATE_RESTARTING, InstanceConstants.STATE_RUNNING))
+                        .and(INSTANCE_HOST_MAP.REMOVED.isNull())
+                        .and(PORT.REMOVED.isNull())
+                        .and(SERVICE_EXPOSE_MAP.UPGRADE.eq(false).or(SERVICE_EXPOSE_MAP.UPGRADE.isNull())))
+                .fetchGroups(INSTANCE_HOST_MAP.HOST_ID);
+
+        Map<Long, List<Port>> hostToPorts = new HashMap<>();
+        for (Map.Entry<Long, Result<Record>> entry : results.entrySet()) {
+            List<Port> ports = new ArrayList<>();
+            hostToPorts.put(entry.getKey(), ports);
+            for (Record rec : entry.getValue()) {
+                PortRecord port = rec.into(PortRecord.class);
+                ports.add(port);
+            }
+        }
+
+        for (CandidateHostInfo hostInfo : hostInfos) {
+            List<Port> ports = hostToPorts.get(hostInfo.getHostId()) != null ? hostToPorts.get(hostInfo.getHostId()) : new ArrayList<Port>();
+            hostInfo.setUsedPorts(ports);
+        }
+    }
+
+    protected SelectConditionStep<Record3<String, Long, Long>> getHostQuery(List<String> orderedHostUUIDs, QueryOptions options) {
+        return create()
+                .select(HOST.UUID, HOST.ID, STORAGE_POOL.ID)
+                .from(HOST)
+                .leftOuterJoin(STORAGE_POOL_HOST_MAP)
+                    .on(STORAGE_POOL_HOST_MAP.HOST_ID.eq(HOST.ID)
+                        .and(STORAGE_POOL_HOST_MAP.REMOVED.isNull()))
+                .join(STORAGE_POOL)
+                    .on(STORAGE_POOL.ID.eq(STORAGE_POOL_HOST_MAP.STORAGE_POOL_ID))
+                .leftOuterJoin(AGENT)
+                    .on(AGENT.ID.eq(HOST.AGENT_ID))
+                .where(getQueryOptionCondition(options, orderedHostUUIDs));
+    }
+
+    protected Condition inHostList(List<String> hostUUIDs) {
+        if (hostUUIDs == null || hostUUIDs.isEmpty()) {
+            return DSL.trueCondition();
+        }
+        return HOST.UUID.in(hostUUIDs);
+    }
+
+    protected Condition getQueryOptionCondition(QueryOptions options, List<String> orderedHostUUIDs) {
+        Condition condition = null;
+
+        if (options.getAccountId() != null) {
+            condition = append(condition, HOST.ACCOUNT_ID.eq(options.getAccountId()));
+        }
+
+        if (options.getRequestedHostId() != null) {
+            condition = append(condition, HOST.ID.eq(options.getRequestedHostId()));
+            return condition;
+        }
+
+        condition = append(condition, AGENT.ID.isNull().or(AGENT.STATE.eq(CommonStatesConstants.ACTIVE))
+        .and(HOST.STATE.in(CommonStatesConstants.ACTIVE, CommonStatesConstants.UPDATING_ACTIVE))
+        .and(STORAGE_POOL.STATE.eq(CommonStatesConstants.ACTIVE))
+        .and(inHostList(orderedHostUUIDs)));
+
+        if ( options.getHosts().size() > 0 ) {
+            condition = append(condition, HOST.ID.in(options.getHosts()));
+        }
+
+        if ( options.getKind() != null ) {
+            condition = append(condition,
+                    HOST.KIND.eq(options.getKind()).and(STORAGE_POOL.KIND.eq(options.getKind())));
+        }
+
+        return condition == null ? DSL.trueCondition() : condition;
+    }
+
+    protected Condition append(Condition base, Condition next) {
+        if ( base == null ) {
+            return next;
+        } else {
+            return base.and(next);
+        }
     }
 }
