+++ code/iaas/logic/pom.xml
@@ -10,11 +10,6 @@
     <dependencies>
         <dependency>
             <groupId>io.cattle</groupId>
-            <artifactId>cattle-iaas-simple-allocator</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>io.cattle</groupId>
             <artifactId>cattle-iaas-agent-instance</artifactId>
             <version>${project.version}</version>
         </dependency>
