+++ code/iaas/service-discovery/server/src/main/java/io/cattle/platform/servicediscovery/service/impl/GlobalHostActivateServiceLookup.java
@@ -7,14 +7,12 @@ import io.cattle.platform.core.model.Service;
 import io.cattle.platform.object.ObjectManager;
 import io.cattle.platform.servicediscovery.api.dao.ServiceDao;
 import io.cattle.platform.servicediscovery.api.dao.ServiceExposeMapDao;
-import io.cattle.platform.servicediscovery.api.util.ServiceDiscoveryUtil;
 import io.cattle.platform.servicediscovery.service.ServiceDiscoveryService;
 import io.cattle.platform.servicediscovery.service.ServiceLookup;
 
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
-import java.util.Map;
 
 import javax.inject.Inject;
 
@@ -49,9 +47,7 @@ public class GlobalHostActivateServiceLookup implements ServiceLookup {
         List<? extends Service> services = expMapDao.getActiveServices(host.getAccountId());
         List<Service> activeGlobalServices = new ArrayList<Service>();
         for (Service service : services) {
-            Map<String, String> serviceLabels = ServiceDiscoveryUtil.getMergedServiceLabels(service, allocationHelper);
-            if ((sdSvc.isGlobalService(service) || sdSvc.isScalePolicyService(service)) &&
-                    allocationHelper.hostChangesAffectsHostAffinityRules(host.getId(), serviceLabels)) {
+            if (sdSvc.isGlobalService(service) || sdSvc.isScalePolicyService(service)) {
                 activeGlobalServices.add(service);
             }
         }
