+++ code/packaging/app-config/pom.xml
@@ -58,11 +58,6 @@
         </dependency>
         <dependency>
             <groupId>io.cattle</groupId>
-            <artifactId>cattle-iaas-simple-allocator</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>io.cattle</groupId>
             <artifactId>cattle-simulator-agent-connection</artifactId>
             <version>${project.version}</version>
         </dependency>
